refactoring | Services complete

+++ src/main/java/ch/zli/m223/service/MemberSercvice.java
@@ -21,7 +21,7 @@ public class MemberSercvice {
     }
 
     @Transactional
-    public void deleteMember(Long id) {
+    public void deleteMember(Integer id) {
         var entity = entityManager.find(Member.class, id);
         entityManager.remove(entity);
     }
